search: factor out lucky search rules and generator
